change qrcodemanager to service
